sw/sublime: change cents to be a 32-bit integer
8-bit isn't enough and 32-bit is used in the internal freq
reg calculations

+++ sw/synth/sublime.h
@@ -22,7 +22,7 @@
 
 struct osc {
 	int enable;
-	int8_t cents;
+	int32_t cents;
 };
 
 struct voice {
